+++ pkg/kubefed2/disable.go
@@ -21,6 +21,7 @@ import (
 	"fmt"
 	"io"
 	"strings"
+	"time"
 
 	"github.com/golang/glog"
 	"github.com/pkg/errors"
@@ -28,6 +29,9 @@ import (
 	"github.com/spf13/pflag"
 
 	apiextv1b1client "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/typed/apiextensions/v1beta1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/rest"
 
 	"github.com/kubernetes-sigs/federation-v2/pkg/apis/core/typeconfig"
@@ -37,6 +41,11 @@ import (
 	"github.com/kubernetes-sigs/federation-v2/pkg/kubefed2/enable"
 	"github.com/kubernetes-sigs/federation-v2/pkg/kubefed2/options"
 	"github.com/kubernetes-sigs/federation-v2/pkg/kubefed2/util"
+)
+
+const (
+	federationGroupUsage = "The name of the API group to use for deleting the federated CRD type when the federated type config does not exist. Only used with --delete-crd."
+	targetVersionUsage   = "The API version of the target type to use for deletion of the federated CRD type when the federated type config does not exist. Only used with --delete-crd."
 )
 
 var (
@@ -55,25 +64,26 @@ var (
 		kubefed2 disable deployments.apps
 
 		# Disable propagation of the kubernetes API type 'Deployment', named
-		in FederatedTypeConfig as 'deployments.apps', and delete corresponding
-		Federated API resources
-		kubefed2 disable deployments.apps --delete-from-api`
+		in FederatedTypeConfig as 'deployments.apps', and delete the
+		corresponding Federated API resource
+		kubefed2 disable deployments.apps --delete-crd`
 )
 
 type disableType struct {
 	options.GlobalSubcommandOptions
+	options.CommonEnableOptions
 	disableTypeOptions
 }
 
 type disableTypeOptions struct {
-	targetName string
-	delete     bool
+	deleteCRD           bool
+	enableTypeDirective *enable.EnableTypeDirective
 }
 
-// Bind adds the join specific arguments to the flagset passed in as an
+// Bind adds the disable specific arguments to the flagset passed in as an
 // argument.
 func (o *disableTypeOptions) Bind(flags *pflag.FlagSet) {
-	flags.BoolVar(&o.delete, "delete-from-api", false, "Whether to remove the API resources added by 'enable'.")
+	flags.BoolVar(&o.deleteCRD, "delete-crd", false, "Whether to remove the API resource added by 'enable'.")
 }
 
 // NewCmdTypeDisable defines the `disable` command that
@@ -101,6 +111,7 @@ func NewCmdTypeDisable(cmdOut io.Writer, config util.FedConfig) *cobra.Command {
 
 	flags := cmd.Flags()
 	opts.GlobalSubcommandBind(flags)
+	opts.CommonSubcommandBind(flags, federationGroupUsage, targetVersionUsage)
 	opts.Bind(flags)
 
 	return cmd
@@ -108,10 +119,27 @@ func NewCmdTypeDisable(cmdOut io.Writer, config util.FedConfig) *cobra.Command {
 
 // Complete ensures that options are valid and marshals them if necessary.
 func (j *disableType) Complete(args []string) error {
-	if len(args) == 0 {
-		return errors.New("NAME is required")
+	j.enableTypeDirective = enable.NewEnableTypeDirective()
+	directive := j.enableTypeDirective
+
+	if err := j.SetName(args); err != nil {
+		return err
 	}
-	j.targetName = args[0]
+
+	if !j.deleteCRD {
+		if len(j.TargetVersion) > 0 {
+			return errors.New("--version flag valid only with --delete-crd")
+		} else if j.FederationGroup != options.DefaultFederationGroup {
+			return errors.New("--federation-group flag valid only with --delete-crd")
+		}
+	}
+
+	if len(j.TargetVersion) > 0 {
+		directive.Spec.TargetVersion = j.TargetVersion
+	}
+	if len(j.FederationGroup) > 0 {
+		directive.Spec.FederationGroup = j.FederationGroup
+	}
 
 	return nil
 }
@@ -126,9 +154,9 @@ func (j *disableType) Run(cmdOut io.Writer, config util.FedConfig) error {
 	// If . is specified, the target name is assumed as a group qualified name.
 	// In such case, ignore the lookup to make sure deletion of a federatedtypeconfig
 	// for which the corresponding target has been removed.
-	name := j.targetName
-	if !strings.Contains(j.targetName, ".") {
-		apiResource, err := enable.LookupAPIResource(hostConfig, j.targetName, "")
+	name := j.TargetName
+	if !strings.Contains(j.TargetName, ".") {
+		apiResource, err := enable.LookupAPIResource(hostConfig, j.TargetName, "")
 		if err != nil {
 			return err
 		}
@@ -139,23 +167,15 @@ func (j *disableType) Run(cmdOut io.Writer, config util.FedConfig) error {
 		Namespace: j.FederationNamespace,
 		Name:      name,
 	}
-
-	return DisableFederation(cmdOut, hostConfig, typeConfigName, j.delete, j.DryRun)
+	j.enableTypeDirective.Name = typeConfigName.Name
+	return DisableFederation(cmdOut, hostConfig, j.enableTypeDirective, typeConfigName, j.deleteCRD, j.DryRun, true)
 }
 
-func DisableFederation(cmdOut io.Writer, config *rest.Config, typeConfigName ctlutil.QualifiedName, delete, dryRun bool) error {
+func DisableFederation(cmdOut io.Writer, config *rest.Config, enableTypeDirective *enable.EnableTypeDirective,
+	typeConfigName ctlutil.QualifiedName, deleteCRD, dryRun, verifyStopped bool) error {
 	client, err := genericclient.New(config)
 	if err != nil {
 		return errors.Wrap(err, "Failed to get federation clientset")
-	}
-	typeConfig := &fedv1a1.FederatedTypeConfig{}
-	err = client.Get(context.TODO(), typeConfig, typeConfigName.Namespace, typeConfigName.Name)
-	if err != nil {
-		return errors.Wrapf(err, "Error retrieving FederatedTypeConfig %q", typeConfigName)
-	}
-
-	if dryRun {
-		return nil
 	}
 
 	write := func(data string) {
@@ -168,9 +188,82 @@ func DisableFederation(cmdOut io.Writer, config *rest.Config, typeConfigName ctl
 		}
 	}
 
+	typeConfig := &fedv1a1.FederatedTypeConfig{}
+	ftcExists, err := checkFederatedTypeConfigExists(client, typeConfig, typeConfigName, write)
+	if err != nil {
+		return err
+	}
+
+	if dryRun {
+		return nil
+	}
+
+	// Disable propagation and verify it is stopped before deleting the CRD
+	// when no custom resources exist. This avoids spurious error messages in
+	// the controller manager log as watches are terminated and cannot be
+	// reestablished.
+	if ftcExists {
+		if deleteCRD {
+			err = checkFederatedTypeCustomResourcesExist(config, typeConfig, write)
+			if err != nil {
+				return err
+			}
+		}
+		if typeConfig.Spec.PropagationEnabled {
+			err = disablePropagation(client, typeConfig, typeConfigName, write)
+			if err != nil {
+				return err
+			}
+		}
+		if verifyStopped {
+			err = verifyPropagationControllerStopped(client, typeConfigName, write)
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	if deleteCRD {
+		if !ftcExists {
+			typeConfig, err = generatedFederatedTypeConfig(config, enableTypeDirective)
+			if err != nil {
+				return err
+			}
+		}
+		err = deleteFederatedType(config, typeConfig, write)
+		if err != nil {
+			return err
+		}
+	}
+
+	if ftcExists {
+		err = deleteFederatedTypeConfig(client, typeConfig, typeConfigName, write)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func checkFederatedTypeConfigExists(client genericclient.Client, typeConfig *fedv1a1.FederatedTypeConfig, typeConfigName ctlutil.QualifiedName, write func(string)) (bool, error) {
+	err := client.Get(context.TODO(), typeConfig, typeConfigName.Namespace, typeConfigName.Name)
+	if err == nil {
+		return true, nil
+	}
+
+	if apierrors.IsNotFound(err) {
+		write(fmt.Sprintf("FederatedTypeConfig %q does not exist\n", typeConfigName))
+		return false, nil
+	}
+
+	return false, errors.Wrapf(err, "Error retrieving FederatedTypeConfig %q", typeConfigName)
+}
+
+func disablePropagation(client genericclient.Client, typeConfig *fedv1a1.FederatedTypeConfig, typeConfigName ctlutil.QualifiedName, write func(string)) error {
 	if typeConfig.Spec.PropagationEnabled {
 		typeConfig.Spec.PropagationEnabled = false
-		err = client.Update(context.TODO(), typeConfig)
+		err := client.Update(context.TODO(), typeConfig)
 		if err != nil {
 			return errors.Wrapf(err, "Error disabling propagation for FederatedTypeConfig %q", typeConfigName)
 		}
@@ -178,36 +271,108 @@ func DisableFederation(cmdOut io.Writer, config *rest.Config, typeConfigName ctl
 	} else {
 		write(fmt.Sprintf("Propagation already disabled for FederatedTypeConfig %q\n", typeConfigName))
 	}
-	if !delete {
-		return nil
-	}
+	return nil
+}
 
-	// TODO(marun) consider waiting for the sync controller to be stopped before attempting deletion
-	err = deleteFederatedType(config, typeConfig, write)
+func verifyPropagationControllerStopped(client genericclient.Client, typeConfigName ctlutil.QualifiedName, write func(string)) error {
+	write(fmt.Sprintf("Verifying propagation controller is stopped for FederatedTypeConfig %q\n", typeConfigName))
+
+	var typeConfig *fedv1a1.FederatedTypeConfig
+	err := wait.PollImmediate(100*time.Millisecond, 10*time.Second, func() (bool, error) {
+		typeConfig = &fedv1a1.FederatedTypeConfig{}
+		err := client.Get(context.TODO(), typeConfig, typeConfigName.Namespace, typeConfigName.Name)
+		if err != nil {
+			glog.Errorf("Error retrieving FederatedTypeConfig %q: %v", typeConfigName, err)
+			return false, nil
+		}
+		if typeConfig.Status.PropagationController == fedv1a1.ControllerStatusNotRunning {
+			return true, nil
+		}
+		return false, nil
+	})
+
 	if err != nil {
-		return err
+		return errors.Wrapf(err, "Unable to verify propagation controller for FederatedTypeConfig %q is stopped: %v", typeConfigName, err)
 	}
 
-	err = client.Delete(context.TODO(), typeConfig, typeConfig.Namespace, typeConfig.Name)
+	write(fmt.Sprintf("Propagation controller for FederatedTypeConfig %q is stopped\n", typeConfigName))
+	return nil
+}
+
+func deleteFederatedTypeConfig(client genericclient.Client, typeConfig *fedv1a1.FederatedTypeConfig, typeConfigName ctlutil.QualifiedName, write func(string)) error {
+	err := client.Delete(context.TODO(), typeConfig, typeConfig.Namespace, typeConfig.Name)
 	if err != nil {
 		return errors.Wrapf(err, "Error deleting FederatedTypeConfig %q", typeConfigName)
 	}
 	write(fmt.Sprintf("federatedtypeconfig %q deleted\n", typeConfigName))
+	return nil
+}
+
+func generatedFederatedTypeConfig(config *rest.Config, enableTypeDirective *enable.EnableTypeDirective) (*fedv1a1.FederatedTypeConfig, error) {
+	apiResource, err := enable.LookupAPIResource(config, enableTypeDirective.Name, enableTypeDirective.Spec.TargetVersion)
+	if err != nil {
+		return nil, err
+	}
+	typeConfig := enable.GenerateTypeConfigForTarget(*apiResource, enableTypeDirective).(*fedv1a1.FederatedTypeConfig)
+	return typeConfig, nil
+}
+
+func deleteFederatedType(config *rest.Config, typeConfig typeconfig.Interface, write func(string)) error {
+	err := checkFederatedTypeCustomResourcesExist(config, typeConfig, write)
+	if err != nil {
+		return err
+	}
+
+	crdName := typeconfig.GroupQualifiedName(typeConfig.GetFederatedType())
+	err = deleteFederatedCRD(config, crdName, write)
+	if err != nil {
+		return err
+	}
 
 	return nil
 }
 
-func deleteFederatedType(config *rest.Config, typeConfig typeconfig.Interface, write func(string)) error {
+func checkFederatedTypeCustomResourcesExist(config *rest.Config, typeConfig typeconfig.Interface, write func(string)) error {
+	federatedTypeAPIResource := typeConfig.GetFederatedType()
+	crdName := typeconfig.GroupQualifiedName(federatedTypeAPIResource)
+	exists, err := customResourcesExist(config, &federatedTypeAPIResource)
+	if err != nil {
+		return err
+	} else if exists {
+		return errors.Errorf("Cannot delete CRD %q while resource instances exist. Please try kubefed2 disable again after removing the resource instances or without the '--delete-crd' option\n", crdName)
+	}
+	return nil
+}
+
+func customResourcesExist(config *rest.Config, resource *metav1.APIResource) (bool, error) {
+	client, err := ctlutil.NewResourceClient(config, resource)
+	if err != nil {
+		return false, err
+	}
+
+	options := metav1.ListOptions{IncludeUninitialized: true}
+	objList, err := client.Resources("").List(options)
+	if apierrors.IsNotFound(err) {
+		return false, nil
+	} else if err != nil {
+		return false, err
+	}
+	return len(objList.Items) != 0, nil
+}
+
+func deleteFederatedCRD(config *rest.Config, crdName string, write func(string)) error {
 	client, err := apiextv1b1client.NewForConfig(config)
 	if err != nil {
 		return errors.Wrap(err, "Error creating crd client")
 	}
 
-	crdName := typeconfig.GroupQualifiedName(typeConfig.GetFederatedType())
 	err = client.CustomResourceDefinitions().Delete(crdName, nil)
-	if err != nil {
+	if apierrors.IsNotFound(err) {
+		write(fmt.Sprintf("customresourcedefinition %q does not exist\n", crdName))
+	} else if err != nil {
 		return errors.Wrapf(err, "Error deleting crd %q", crdName)
+	} else {
+		write(fmt.Sprintf("customresourcedefinition %q deleted\n", crdName))
 	}
-	write(fmt.Sprintf("customresourcedefinition %q deleted\n", crdName))
 	return nil
 }