+++ pkg/kubefed2/options/options.go
@@ -48,16 +48,16 @@ func (o *GlobalSubcommandOptions) GlobalSubcommandBind(flags *pflag.FlagSet) {
 		"Run the command in dry-run mode, without making any server requests.")
 }
 
-// CommonSubcommandOptions holds the common configuration required by some of
-// the subcommands of `kubefed2`.
-type CommonSubcommandOptions struct {
+// CommonJoinOptions holds the common configuration required by the join and
+// unjoin subcommands of `kubefed2`.
+type CommonJoinOptions struct {
 	ClusterName     string
 	ClusterContext  string
 	HostClusterName string
 }
 
 // CommonSubcommandBind adds the common subcommand flags to the flagset passed in.
-func (o *CommonSubcommandOptions) CommonSubcommandBind(flags *pflag.FlagSet) {
+func (o *CommonJoinOptions) CommonSubcommandBind(flags *pflag.FlagSet) {
 	flags.StringVar(&o.ClusterContext, "cluster-context", "",
 		"Name of the cluster's context in the local kubeconfig. Defaults to cluster name if unspecified.")
 	flags.StringVar(&o.HostClusterName, "host-cluster-name", "",
@@ -66,7 +66,7 @@ func (o *CommonSubcommandOptions) CommonSubcommandBind(flags *pflag.FlagSet) {
 
 // SetName sets the name from the args passed in for the required positional
 // argument.
-func (o *CommonSubcommandOptions) SetName(args []string) error {
+func (o *CommonJoinOptions) SetName(args []string) error {
 	if len(args) == 0 {
 		return errors.New("NAME is required")
 	}
@@ -98,4 +98,36 @@ func GetOptionsFromFederationConfig(hostConfig *rest.Config, namespace string) (
 	}
 
 	return fedConfig.Spec.Scope, fedConfig.Spec.RegistryNamespace, nil
+}
+
+// CommonEnableOptions holds the common configuration required by the enable
+// and disable subcommands of `kubefed2`.
+type CommonEnableOptions struct {
+	TargetName      string
+	FederationGroup string
+	TargetVersion   string
+}
+
+// Default value for shared Federation group across enable and
+// disable subcommands of `kubefed2`.
+const (
+	DefaultFederationGroup   = "types.federation.k8s.io"
+	DefaultFederationVersion = "v1alpha1"
+)
+
+// CommonSubcommandBind adds the common subcommand flags to the flagset passed in.
+func (o *CommonEnableOptions) CommonSubcommandBind(flags *pflag.FlagSet, federationGroupUsage, targetVersionUsage string) {
+	flags.StringVar(&o.FederationGroup, "federation-group", DefaultFederationGroup, federationGroupUsage)
+	flags.StringVar(&o.TargetVersion, "version", "", targetVersionUsage)
+}
+
+// SetName sets the name from the args passed in for the required positional
+// argument.
+func (o *CommonEnableOptions) SetName(args []string) error {
+	if len(args) == 0 {
+		return errors.New("NAME is required")
+	}
+
+	o.TargetName = args[0]
+	return nil
 }